Modified get ignore to not track .log files. I also re-factored and Modularized Java files
