+++ .gitignore
@@ -59,6 +59,7 @@ build/
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+*.log
 
 # gitignore for Node express
 /node_app/node_modules
