--- node_app/Middleware/debug.log
@@ -1,1 +1,0 @@
-recipes::addRecipe::success
