+++ recipe_box/src/main/java/com/spr/recipe_box/Constant/Constants.java
@@ -1,13 +1,15 @@
-package com.spr.recipe_box.Constants;
+package com.spr.recipe_box.Constant;
 
 public class Constants {
     public static final String DEV_BOOL = "DEV";
-    public static final String RECIPES = "recipes/";
     public static final String RECIPE = "recipe/";
+    public static final String RECIPES = "recipes/";
     public static final String USER = "user/";
+    public static final String TOP25 = "top25random";
     public static final String API_KEY_QUERY = "?api_key=";
     public static final String USER_NAME_QUERY = "user?user_name=";
     public static final String NODE_DEV_ENV = "http://127.0.0.1/";
     public static final String NODE_PROD_ENV = "http://13.56.134.63/";
     public static final String BIG_OVEN = "https://api.bigoven.com/";
+    public static final String BIG_OVEN_API_2 = "https://api2.bigoven.com/";
 }
