+++ recipe_box/src/main/java/com/spr/recipe_box/Constant/Keys.java
@@ -1,0 +1,5 @@
+package com.spr.recipe_box.Constant;
+
+public class Keys {
+    public static final String BIG_OVEN_KEY = "glFUKikehWjLW900etpS564VgIzOWSW5";
+}
