+++ recipe_box/src/main/java/com/spr/recipe_box/Controller/RecipeAPIRestController.java
@@ -1,5 +1,7 @@
-package com.spr.recipe_box;
+package com.spr.recipe_box.Controller;
 
+import com.spr.recipe_box.Model.Recipe;
+import com.spr.recipe_box.Service.RecipeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
