+++ recipe_box/src/main/java/com/spr/recipe_box/Controller/RecipeDBRestController.java
@@ -1,6 +1,7 @@
-package com.spr.recipe_box;
+package com.spr.recipe_box.Controller;
 
-import com.spr.recipe_box.Constants.Constants;
+import com.spr.recipe_box.Constant.Constants;
+import com.spr.recipe_box.Config.RestClass;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
