+++ recipe_box/src/main/java/com/spr/recipe_box/Controller/UserRestController.java
@@ -1,8 +1,9 @@
-package com.spr.recipe_box;
+package com.spr.recipe_box.Controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.spr.recipe_box.Constants.Constants;
+import com.spr.recipe_box.Constant.Constants;
+import com.spr.recipe_box.Config.RestClass;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
